+++ include/franka/robot_state.h
@@ -369,6 +369,20 @@ struct RobotState {
   std::array<double, 7> dtheta{};
 
   /**
+   * \f$\ddot{x}\f$
+   * Translational acceleration data from the accelerometers located on the top of the PCB
+   * of Joint 1 through Joint 6 (no accelerometers in the last joint). Unit: \f$[\frac{m}{s^2}]\f$
+   */
+  std::array<std::array<float, 3>, 6> accelerometer_top{};
+
+  /**
+   * \f$\ddot{x}\f$
+   * Translational acceleration data from the accelerometers located on the bottom of the PCB
+   * of Joint 1 through Joint 6 (no accelerometers in the last joint). Unit: \f$[\frac{m}{s^2}]\f$
+   */
+  std::array<std::array<float, 3>, 6> accelerometer_bottom{};
+
+  /**
    * Current error state.
    */
   Errors current_errors{};